Type Alias wasmtime_wasi_http::bindings::wasi::http::outgoing_handler::ErrorCode

pub type ErrorCode = ErrorCode;

Aliased Type

enum ErrorCode {

    DnsTimeout,
    DnsError(DnsErrorPayload),
    DestinationNotFound,
    DestinationUnavailable,
    DestinationIpProhibited,
    DestinationIpUnroutable,
    ConnectionRefused,
    ConnectionTerminated,
    ConnectionTimeout,
    ConnectionReadTimeout,
    ConnectionWriteTimeout,
    ConnectionLimitReached,
    TlsProtocolError,
    TlsCertificateError,
    TlsAlertReceived(TlsAlertReceivedPayload),
    HttpRequestDenied,
    HttpRequestLengthRequired,
    HttpRequestBodySize(Option<u64>),
    HttpRequestMethodInvalid,
    HttpRequestUriInvalid,
    HttpRequestUriTooLong,
    HttpRequestHeaderSectionSize(Option<u32>),
    HttpRequestHeaderSize(Option<FieldSizePayload>),
    HttpRequestTrailerSectionSize(Option<u32>),
    HttpRequestTrailerSize(FieldSizePayload),
    HttpResponseIncomplete,
    HttpResponseHeaderSectionSize(Option<u32>),
    HttpResponseHeaderSize(FieldSizePayload),
    HttpResponseBodySize(Option<u64>),
    HttpResponseTrailerSectionSize(Option<u32>),
    HttpResponseTrailerSize(FieldSizePayload),
    HttpResponseTransferCoding(Option<String>),
    HttpResponseContentCoding(Option<String>),
    HttpResponseTimeout,
    HttpUpgradeFailed,
    HttpProtocolError,
    LoopDetected,
    ConfigurationError,
    InternalError(Option<String>),

}

Variants

DnsTimeout

DnsError(DnsErrorPayload)

DestinationNotFound

DestinationUnavailable

DestinationIpProhibited

DestinationIpUnroutable

ConnectionRefused

ConnectionTerminated

ConnectionTimeout

ConnectionReadTimeout

ConnectionWriteTimeout

ConnectionLimitReached

TlsProtocolError

TlsCertificateError

TlsAlertReceived(TlsAlertReceivedPayload)

HttpRequestDenied

HttpRequestLengthRequired

HttpRequestBodySize(Option<u64>)

HttpRequestMethodInvalid

HttpRequestUriInvalid

HttpRequestUriTooLong

HttpRequestHeaderSectionSize(Option<u32>)

HttpRequestHeaderSize(Option<FieldSizePayload>)

HttpRequestTrailerSectionSize(Option<u32>)

HttpRequestTrailerSize(FieldSizePayload)

HttpResponseIncomplete

HttpResponseHeaderSectionSize(Option<u32>)

HttpResponseHeaderSize(FieldSizePayload)

HttpResponseBodySize(Option<u64>)

HttpResponseTrailerSectionSize(Option<u32>)

HttpResponseTrailerSize(FieldSizePayload)

HttpResponseTransferCoding(Option<String>)

HttpResponseContentCoding(Option<String>)

HttpResponseTimeout

HttpUpgradeFailed

HttpProtocolError

LoopDetected

ConfigurationError

InternalError(Option<String>)

This is a catch-all error for anything that doesnโt fit cleanly into a more specific case. It also includes an optional string for an unstructured description of the error. Users should not depend on the string for diagnosing errors, as itโs not required to be consistent between implementations.