Module wasmtime_wasi_http::bindings::wasi::http::types

Re-exports

pub use super::super::super::__with_name0 as Fields;

pub use super::super::super::__with_name4 as IncomingRequest;

pub use super::super::super::__with_name7 as OutgoingRequest;

pub use super::super::super::__with_name9 as RequestOptions;

pub use super::super::super::__with_name10 as ResponseOutparam;

pub use super::super::super::__with_name5 as IncomingResponse;

pub use super::super::super::__with_name3 as IncomingBody;

pub use super::super::super::__with_name2 as FutureTrailers;

pub use super::super::super::__with_name8 as OutgoingResponse;

pub use super::super::super::__with_name6 as OutgoingBody;

pub use super::super::super::__with_name1 as FutureIncomingResponse;

Structs

DnsErrorPayload

Defines the case payload type for DNS-error above:

FieldSizePayload

Defines the case payload type for HTTP-response-{header,trailer}-size above:

TlsAlertReceivedPayload

Defines the case payload type for TLS-alert-received above:

Enums

ErrorCode

These cases are inspired by the IANA HTTP Proxy Error Types: https://www.iana.org/assignments/http-proxy-status/http-proxy-status.xhtml#table-http-proxy-error-types

HeaderError

This type enumerates the different kinds of errors that may occur when setting or appending to a fields resource.

Method

This type corresponds to HTTP standard Methods.

Scheme

This type corresponds to HTTP standard Related Schemes.

Traits

Host

HostFields

HostFutureIncomingResponse

HostFutureTrailers

HostIncomingBody

HostIncomingRequest

HostIncomingResponse

HostOutgoingBody

HostOutgoingRequest

HostOutgoingResponse

HostRequestOptions

HostResponseOutparam

Functions

add_to_linker

Type Aliases

Duration

FieldKey

Field keys are always strings.

FieldValue

Field values should always be ASCII strings. However, in reality, HTTP implementations often have to interpret malformed values, so they are provided as a list of bytes.

Headers

Headers is an alias for Fields.

InputStream

IoError

OutputStream

Pollable

StatusCode

This type corresponds to the HTTP standard Status Code.

Trailers

Trailers is an alias for Fields.