Trait wasmtime_wasi_http::bindings::wasi::http::types::HostFields

pub trait HostFields {
    // Required methods
    fn new(&mut self) -> Result<Resource<Fields>>;
    fn from_list(
        &mut self,
        entries: Vec<(FieldKey, FieldValue)>
    ) -> Result<Result<Resource<Fields>, HeaderError>>;
    fn get(
        &mut self,
        self_: Resource<Fields>,
        name: FieldKey
    ) -> Result<Vec<FieldValue>>;
    fn set(
        &mut self,
        self_: Resource<Fields>,
        name: FieldKey,
        value: Vec<FieldValue>
    ) -> Result<Result<(), HeaderError>>;
    fn delete(
        &mut self,
        self_: Resource<Fields>,
        name: FieldKey
    ) -> Result<Result<(), HeaderError>>;
    fn append(
        &mut self,
        self_: Resource<Fields>,
        name: FieldKey,
        value: FieldValue
    ) -> Result<Result<(), HeaderError>>;
    fn entries(
        &mut self,
        self_: Resource<Fields>
    ) -> Result<Vec<(FieldKey, FieldValue)>>;
    fn clone(&mut self, self_: Resource<Fields>) -> Result<Resource<Fields>>;
    fn drop(&mut self, rep: Resource<Fields>) -> Result<()>;
}

Required Methods

fn new(&mut self) -> Result<Resource<Fields>>

Construct an empty HTTP Fields.

The resulting fields is mutable.

fn from_list( &mut self, entries: Vec<(FieldKey, FieldValue)> ) -> Result<Result<Resource<Fields>, HeaderError>>

Construct an HTTP Fields.

The resulting fields is mutable.

The list represents each key-value pair in the Fields. Keys which have multiple values are represented by multiple entries in this list with the same key.

The tuple is a pair of the field key, represented as a string, and Value, represented as a list of bytes. In a valid Fields, all keys and values are valid UTF-8 strings. However, values are not always well-formed, so they are represented as a raw list of bytes.

An error result will be returned if any header or value was syntactically invalid, or if a header was forbidden.

fn get( &mut self, self_: Resource<Fields>, name: FieldKey ) -> Result<Vec<FieldValue>>

Get all of the values corresponding to a key.

fn set( &mut self, self_: Resource<Fields>, name: FieldKey, value: Vec<FieldValue> ) -> Result<Result<(), HeaderError>>

Set all of the values for a key. Clears any existing values for that key, if they have been set.

Fails with header-error.immutable if the fields are immutable.

fn delete( &mut self, self_: Resource<Fields>, name: FieldKey ) -> Result<Result<(), HeaderError>>

Delete all values for a key. Does nothing if no values for the key exist.

Fails with header-error.immutable if the fields are immutable.

fn append( &mut self, self_: Resource<Fields>, name: FieldKey, value: FieldValue ) -> Result<Result<(), HeaderError>>

Append a value for a key. Does not change or delete any existing values for that key.

Fails with header-error.immutable if the fields are immutable.

fn entries( &mut self, self_: Resource<Fields> ) -> Result<Vec<(FieldKey, FieldValue)>>

Retrieve the full set of keys and values in the Fields. Like the constructor, the list represents each key-value pair.

The outer list represents each key-value pair in the Fields. Keys which have multiple values are represented by multiple entries in this list with the same key.

fn clone(&mut self, self_: Resource<Fields>) -> Result<Resource<Fields>>

Make a deep copy of the Fields. Equivelant in behavior to calling the fields constructor on the return value of entries. The resulting fields is mutable.

fn drop(&mut self, rep: Resource<Fields>) -> Result<()>