Trait wasmtime_wasi_http::bindings::wasi::http::types::HostRequestOptions

pub trait HostRequestOptions {
    // Required methods
    fn new(&mut self) -> Result<Resource<RequestOptions>>;
    fn connect_timeout_ms(
        &mut self,
        self_: Resource<RequestOptions>
    ) -> Result<Option<Duration>>;
    fn set_connect_timeout_ms(
        &mut self,
        self_: Resource<RequestOptions>,
        ms: Option<Duration>
    ) -> Result<Result<(), ()>>;
    fn first_byte_timeout_ms(
        &mut self,
        self_: Resource<RequestOptions>
    ) -> Result<Option<Duration>>;
    fn set_first_byte_timeout_ms(
        &mut self,
        self_: Resource<RequestOptions>,
        ms: Option<Duration>
    ) -> Result<Result<(), ()>>;
    fn between_bytes_timeout_ms(
        &mut self,
        self_: Resource<RequestOptions>
    ) -> Result<Option<Duration>>;
    fn set_between_bytes_timeout_ms(
        &mut self,
        self_: Resource<RequestOptions>,
        ms: Option<Duration>
    ) -> Result<Result<(), ()>>;
    fn drop(&mut self, rep: Resource<RequestOptions>) -> Result<()>;
}

Required Methods

fn new(&mut self) -> Result<Resource<RequestOptions>>

Construct a default request-options value.

fn connect_timeout_ms( &mut self, self_: Resource<RequestOptions> ) -> Result<Option<Duration>>

The timeout for the initial connect to the HTTP Server.

fn set_connect_timeout_ms( &mut self, self_: Resource<RequestOptions>, ms: Option<Duration> ) -> Result<Result<(), ()>>

Set the timeout for the initial connect to the HTTP Server. An error return value indicates that this timeout is not supported.

fn first_byte_timeout_ms( &mut self, self_: Resource<RequestOptions> ) -> Result<Option<Duration>>

The timeout for receiving the first byte of the Response body.

fn set_first_byte_timeout_ms( &mut self, self_: Resource<RequestOptions>, ms: Option<Duration> ) -> Result<Result<(), ()>>

Set the timeout for receiving the first byte of the Response body. An error return value indicates that this timeout is not supported.

fn between_bytes_timeout_ms( &mut self, self_: Resource<RequestOptions> ) -> Result<Option<Duration>>

The timeout for receiving subsequent chunks of bytes in the Response body stream.

fn set_between_bytes_timeout_ms( &mut self, self_: Resource<RequestOptions>, ms: Option<Duration> ) -> Result<Result<(), ()>>

Set the timeout for receiving subsequent chunks of bytes in the Response body stream. An error return value indicates that this timeout is not supported.

fn drop(&mut self, rep: Resource<RequestOptions>) -> Result<()>