Trait wasmtime_wasi_http::proxy::wasi::io::error::HostError

pub trait HostError {
    // Required methods
    fn to_debug_string<'life0, 'async_trait>(
        &'life0 mut self,
        self_: Resource<Error>
    ) -> Pin<Box<dyn Future<Output = Result<String>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn drop(&mut self, rep: Resource<Error>) -> Result<()>;
}

Required Methods

fn to_debug_string<'life0, 'async_trait>( &'life0 mut self, self_: Resource<Error> ) -> Pin<Box<dyn Future<Output = Result<String>> + Send + 'async_trait>>where Self: 'async_trait, 'life0: 'async_trait,

Returns a string that is suitable to assist humans in debugging this error.

WARNING: The returned string should not be consumed mechanically! It may change across platforms, hosts, or other implementation details. Parsing this string is a major platform-compatibility hazard.

fn drop(&mut self, rep: Resource<Error>) -> Result<()>