Trait wasmtime_wasi_http::bindings::wasi::clocks::monotonic_clock::Host

pub trait Host {
    // Required methods
    fn now(&mut self) -> Result<Instant>;
    fn resolution(&mut self) -> Result<Duration>;
    fn subscribe_instant(&mut self, when: Instant) -> Result<Resource<Pollable>>;
    fn subscribe_duration(
        &mut self,
        when: Duration
    ) -> Result<Resource<Pollable>>;
}

Required Methods

fn now(&mut self) -> Result<Instant>

Read the current value of the clock.

The clock is monotonic, therefore calling this function repeatedly will produce a sequence of non-decreasing values.

fn resolution(&mut self) -> Result<Duration>

Query the resolution of the clock. Returns the duration of time corresponding to a clock tick.

fn subscribe_instant(&mut self, when: Instant) -> Result<Resource<Pollable>>

Create a pollable which will resolve once the specified instant occured.

fn subscribe_duration(&mut self, when: Duration) -> Result<Resource<Pollable>>

Create a pollable which will resolve once the given duration has elapsed, starting at the time at which this function was called. occured.